requirements.txt
PyQt5
SQLAlchemy
cppimport
ipdb
jupyter
numba
orange3
pandasql
pomegranate
psycopg2
pybind11
pygraphviz
python-language-server[all]
scikit-learn
seaborn
tensorflow
toolz
tqdm
